requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
anthropic==0.43.0
slack-sdk==3.34.0
aiohttp==3.11.11
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2
python-dotenv==1.0.1
